+++ src/utils/functions.ts
@@ -1,7 +1,13 @@
+import { randomBytes } from 'crypto';
+
 export function logError<T>(data: T): any {
   return process.stdout.write(`\x1b[31mError PID ${process.pid}:\x1b[0m  ${data}\n`);
 }
 
 export function isFunction<T>(fn: T): boolean {
   return {}.toString.call(fn) === '[object Function]';
+}
+
+export function generateKey(length: number): string {
+  return randomBytes(length).toString('hex');
 }